Change temperature parameter type to float in LCD task and format output to one decimal place in max/min task

+++ tasks/lcd/lcd_task.cpp
@@ -54,7 +54,7 @@ void vLCDTask(void *pvParameters) {
       }
       break;
     case Temperature:
-      int xTemperature = xMessage.xLCDData.xTemperature;
+      float xTemperature = xMessage.xLCDData.xTemperature;
       write_temperature(xTemperature);
       break;
     case BubbleLevel:
